python_ai_service/pyproject.toml
[project]
name = "tvwallau-ai-service"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic",
  "python-dotenv",
  "numpy",
  "requests",
  "Pillow",
  "filelock>=3.13.0",
  "onnx>=1.16.0",
  "onnxscript>=0.1.0",
  "openvino==2024.4.0",
  "openvino-genai==2024.4.0",
  "openvino-tokenizers==2024.4.0",
  "torch",
  "datamodel-code-generator",
  # FEHLT für MODEL_FETCH_MODE=download:
  "optimum-intel[openvino]",
  "transformers",
  "huggingface-hub",
  # optional, aber häufig nötig:
  "sentencepiece",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

# Wichtig: verhindert, dass setuptools aus Versehen "target" als Package mitsammelt
[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["target*", ".venv*", "scripts*", "tests*"]

[dependency-groups]
dev = ["pytest"]
